+++ c-implementation/tab.c
@@ -11,7 +11,7 @@
 #include <stdint.h> // uint8_t, uint32_t
 
 // TAB
-#include <tab.h>    // Header file
+#include <tab.h>    // TAB header
 
 // External handler functions
 
@@ -209,7 +209,6 @@ void write_reply(rx_cmd_buff_t* rx_cmd_buff_o, tx_cmd_buff_t* tx_cmd_buff_o) {
         break;
       case BOOTLOADER_WRITE_PAGE_OPCODE:
         if(bootloader_active()) {
-          // initialize common variables to known values
           success = handle_bootloader_write_page(rx_cmd_buff_o);
           if(success) {
             tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x07);
@@ -227,7 +226,6 @@ void write_reply(rx_cmd_buff_t* rx_cmd_buff_o, tx_cmd_buff_t* tx_cmd_buff_o) {
         break;
       case BOOTLOADER_WRITE_PAGE_ADDR32_OPCODE:
         if(bootloader_active()) {
-          // initialize common variables to known values
           success = handle_bootloader_write_page_addr32(rx_cmd_buff_o);
           if(success) {
             tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x0a);
@@ -255,7 +253,7 @@ void write_reply(rx_cmd_buff_t* rx_cmd_buff_o, tx_cmd_buff_t* tx_cmd_buff_o) {
           if(success) {
             tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x07);
             tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_ACK_OPCODE;
-            tx_cmd_buff_o->data[PLD_START_INDEX] = BOOTLOADER_ACK_REASON_JUMP;
+            tx_cmd_buff_o->data[PLD_START_INDEX] = BOOTLOADER_ACK_REASON_JUMPED;
           } else {
             tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
             tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_NACK_OPCODE;